refactor input excel model types
